backend/requirements.txt
Flask==2.3.3
openai
opencv-python
pytesseract
python-dotenv
flask-mailman
werkzeug
jwt
psycopg2
redis
gunicorn